refactor: auth 패키지 내부 도메인 기반으로 정리
