+++ backend/api/src/main/java/com/yat2/episode/auth/AuthController.java
@@ -1,11 +1,11 @@
 package com.yat2.episode.auth;
 
-import com.yat2.episode.auth.config.AuthRedirectProperties;
-import com.yat2.episode.auth.dto.IssuedTokens;
-import com.yat2.episode.auth.config.KakaoProperties;
+import com.yat2.episode.auth.oauth.OAuthRedirectProperties;
+import com.yat2.episode.auth.jwt.IssuedTokens;
+import com.yat2.episode.auth.oauth.KakaoProperties;
 import com.yat2.episode.auth.oauth.OAuthUtil;
 import com.yat2.episode.auth.refresh.RefreshTokenService;
-import com.yat2.episode.auth.token.AuthCookieFactory;
+import com.yat2.episode.auth.cookie.AuthCookieFactory;
 import com.yat2.episode.global.exception.CustomException;
 import com.yat2.episode.global.exception.ErrorCode;
 import io.swagger.v3.oas.annotations.Operation;
@@ -35,7 +35,7 @@ public class AuthController {
     private final KakaoProperties kakaoProperties;
     private final AuthService authService;
     private final AuthCookieFactory authCookieFactory;
-    private final AuthRedirectProperties authRedirectProperties;
+    private final OAuthRedirectProperties OAuthRedirectProperties;
     private final RefreshTokenService refreshTokenService;
 
     @GetMapping("/login")
@@ -108,7 +108,7 @@ public class AuthController {
         ResponseCookie accessCookie = authCookieFactory.access(tokens.accessToken());
         ResponseCookie refreshCookie = authCookieFactory.refresh(tokens.refreshToken());
 
-        String redirect = isLocalDev ? authRedirectProperties.getLocal() : authRedirectProperties.getProd();
+        String redirect = isLocalDev ? OAuthRedirectProperties.getLocal() : OAuthRedirectProperties.getProd();
 
         return ResponseEntity.status(302)
                 .header(HttpHeaders.LOCATION, redirect)
