+++ backend/api/src/main/java/com/yat2/episode/auth/AuthService.java
@@ -1,10 +1,10 @@
 package com.yat2.episode.auth;
 
 import com.nimbusds.jwt.JWTClaimsSet;
-import com.yat2.episode.auth.dto.IssuedTokens;
-import com.yat2.episode.auth.dto.KakaoTokenResponse;
+import com.yat2.episode.auth.jwt.IssuedTokens;
+import com.yat2.episode.auth.oauth.KakaoTokenResponse;
 import com.yat2.episode.auth.refresh.RefreshTokenService;
-import com.yat2.episode.auth.token.JwtProvider;
+import com.yat2.episode.auth.jwt.JwtProvider;
 import com.yat2.episode.auth.oauth.KakaoIdTokenVerifier;
 import com.yat2.episode.auth.oauth.KakaoOAuthClient;
 import com.yat2.episode.users.Users;
