+++ backend/api/src/main/java/com/yat2/episode/auth/cookie/AuthCookieNames.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.auth.token;
+package com.yat2.episode.auth.cookie;
 
 public final class AuthCookieNames {
     private AuthCookieNames() {}
