+++ backend/api/src/main/java/com/yat2/episode/auth/jwt/IssuedTokens.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.auth.dto;
+package com.yat2.episode.auth.jwt;
 
 public record IssuedTokens(
         String accessToken,
