+++ backend/api/src/main/java/com/yat2/episode/auth/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.auth.config;
+package com.yat2.episode.auth.jwt;
 
 import lombok.Getter;
 import lombok.Setter;
