+++ backend/api/src/main/java/com/yat2/episode/auth/jwt/JwtProvider.java
@@ -1,12 +1,10 @@
-package com.yat2.episode.auth.token;
+package com.yat2.episode.auth.jwt;
 
 import com.nimbusds.jose.*;
 import com.nimbusds.jose.crypto.MACSigner;
 import com.nimbusds.jose.crypto.MACVerifier;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
-import com.yat2.episode.auth.config.JwtProperties;
-import com.yat2.episode.auth.dto.IssuedTokens;
 import com.yat2.episode.global.exception.CustomException;
 import com.yat2.episode.global.exception.ErrorCode;
 import org.springframework.stereotype.Component;
