+++ backend/api/src/main/java/com/yat2/episode/auth/oauth/OAuthRedirectProperties.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.auth.config;
+package com.yat2.episode.auth.oauth;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,7 +6,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "auth.redirect")
 @Getter @Setter
-public class AuthRedirectProperties {
+public class OAuthRedirectProperties {
     private String local;
     private String prod;
 }
